refactor: suppression creation
