+++ client/src/components/molecules/Creation.jsx
@@ -16,7 +16,7 @@ const Wrapper = styled.div`
     font-family: "Ruluko", Arial, Sans-serif;
 `
 
-const Suprime = styled.div`
+const Supprime = styled.div`
     display: flex;
     justify-content: flex-end;
 `
@@ -31,10 +31,10 @@ class Creation extends Component {
     }
 
     handleDeleteClick = async () => {
-        if (await deleteCreation(this.props.valueId))
+        if (await deleteCreation(this.props.valueId)) {
             this.props.msgAction("Supression effectuée avec succès")
-        else this.props.msgAction("Erreur dans la suppression")
-        this.setState({ redirect: <Redirect to="/accueil" /> })
+            this.setState({ redirect: <Redirect to="/accueil" /> })
+        } else this.props.msgAction("Erreur dans la suppression")
     }
 
     render() {
@@ -49,15 +49,15 @@ class Creation extends Component {
 
                     <Wrapper>
                         <div>{this.props.description}</div>
-                        <Suprime>
+                        <Supprime>
                             <Link to={"/updateCreation/" + this.props.valueId}>
                                 Modifier
                             </Link>
                             <button
-                                className="far fa-times-circle fa-2x"
+                                className="far fa-times-circle fa-2x deleteButton"
                                 onClick={this.handleDeleteClick}
                             />
-                        </Suprime>
+                        </Supprime>
                     </Wrapper>
                     {this.state.redirect}
                 </React.Fragment>
